better prefetch_checksums.sh

+++ prefetch_checksums.sh
@@ -1,5 +1,12 @@
 #!/run/current-system/sw/bin/bash
 
+# Color codes
+RED="\e[31m"
+GREEN="\e[32m"
+YELLOW="\e[33m"
+LIGHT_GRAY="\e[37m"
+RESET="\e[0m"
+
 # Check if the number of arguments is not 0 or 2
 if [[ $# -ne 0 && $# -ne 2 ]]; then
     echo "Usage:"
@@ -8,10 +15,15 @@ if [[ $# -ne 0 && $# -ne 2 ]]; then
     exit 1
 fi
 
-# Versions
+# Metadata line range
+metadata_begin_line=12
+metadata_end_line=16
+
+# Requested versions
 iosevka_version=""
 nerdfontpatcher_version=""
 if [[ $# -eq 0 ]]; then
+  echo -e "${LIGHT_GRAY}Fetching latest versions...${RESET}"
   iosevka_version_tag=$(curl -s https://api.github.com/repos/be5invis/Iosevka/releases/latest | jq -r '.tag_name')
   nerdfontpatcher_version_tag=$(curl -s https://api.github.com/repos/ryanoasis/nerd-fonts/releases/latest | jq -r '.tag_name')
   iosevka_version=${iosevka_version_tag:1}
@@ -22,22 +34,46 @@ if [[ $# -eq 2 ]]; then
   nerdfontpatcher_version="$2"
 fi
 
-echo "Iosevka: $iosevka_version"
-echo "nerd-fonts: $nerdfontpatcher_version"
+# Current versions
+current_iosevka_version=$(cat flake.nix | grep -oE 'version = "[0-9]+\.[0-9]+\.[0-9]+"' | grep -oE '[0-9]+\.[0-9]+\.[0-9]+')
+current_nerdfontpatcher_version=$(cat flake.nix | grep -oE 'fontPatcherVersion = "[0-9]+\.[0-9]+\.[0-9]+"' | grep -oE '[0-9]+\.[0-9]+\.[0-9]+')
 
-# Direct checksums
-# iosevka_checksum=$(nix-prefetch --option extra-experimental-features flakes fetchFromGitHub --owner be5invis --repo iosevka --rev "v$iosevka_version" --check-store)
-iosevka_checksum=$(nix-prefetch --option extra-experimental-features flakes fetchzip --url "https://github.com/be5invis/Iosevka/archive/refs/tags/v$iosevka_version.zip" --check-store)
-nerdfontpatcher_checksum=$(nix-prefetch --option extra-experimental-features flakes fetchzip --url "https://github.com/ryanoasis/nerd-fonts/releases/download/v$nerdfontpatcher_version/FontPatcher.zip" --no-stripRoot --check-store)
+# Compare versions
+if [[ "$current_iosevka_version" != "$iosevka_version" ]]; then
+  iosevka_needs_update=true
+  echo -e "Iosevka: ${RED}$current_iosevka_version${RESET} -> ${GREEN}$iosevka_version${RESET}"
+else
+  echo "Iosevka: $current_iosevka_version -> $iosevka_version"
+fi
 
-# NPM dependencies checksum
+if [[ "$current_nerdfontpatcher_version" != "$nerdfontpatcher_version" ]]; then
+  nerdfontpatcher_needs_update=true
+  echo -e "nerd-fonts: ${RED}$current_nerdfontpatcher_version${RESET} -> ${GREEN}$nerdfontpatcher_version${RESET}"
+else
+  echo "nerd-fonts: $current_nerdfontpatcher_version -> $nerdfontpatcher_version"
+fi
+
+if [[ "$iosevka_needs_update" != true && "$nerdfontpatcher_needs_update" != true ]]; then
+  echo -e "${YELLOW}Nothing to update, exiting without any change to flake.nix${RESET}"
+  exit 0
+fi
+
+# Direct hashes
+# remove --silent for debugging
+echo -e "${LIGHT_GRAY}Calculating hashes...${RESET}"
+# iosevka_checksum=$(nix-prefetch --option extra-experimental-features flakes fetchFromGitHub --owner be5invis --repo iosevka --rev "v$iosevka_version" --check-store --silent)
+iosevka_checksum=$(nix-prefetch --option extra-experimental-features flakes fetchzip --url "https://github.com/be5invis/Iosevka/archive/refs/tags/v$iosevka_version.zip" --check-store --silent)
+nerdfontpatcher_checksum=$(nix-prefetch --option extra-experimental-features flakes fetchzip --url "https://github.com/ryanoasis/nerd-fonts/releases/download/v$nerdfontpatcher_version/FontPatcher.zip" --no-stripRoot --check-store --silent)
+
+# NPM dependencies hash
+echo -e "${LIGHT_GRAY}Calculating dependency (NPM packages) hashes...${RESET}"
 filename=$(mktemp)
 curl -s -L "https://raw.githubusercontent.com/be5invis/Iosevka/v$iosevka_version/package-lock.json" -o "$filename"
 iosevka_npmdeps_checksum=$(prefetch-npm-deps "$filename")
 rm "$filename"
 
-# Output
-formatted_output=$(cat << EOF
+# Updated hashes
+updated_hashes=$(cat << EOF
     version = "$iosevka_version";
     hash = "$iosevka_checksum";
     npmDepsHash = "$iosevka_npmdeps_checksum";
@@ -46,29 +82,31 @@ formatted_output=$(cat << EOF
 EOF
 )
 
-echo -e "\033[1;33m------ Ignore the output above ------\033[0m"
+# Original hashes
+original_hashes=$(sed -n "${metadata_begin_line},${metadata_end_line}p" flake.nix)
 
-echo -e "\033[1;33m------ Prefetched hashes ------\033[0m"
-echo -e "\033[1;32m$formatted_output\033[0m"
-echo -e "\033[1;33m------ Prefetched hashes ------\033[0m"
+# Show diff with difftastic
+echo -e "${LIGHT_GRAY}Changes to flake.nix${RESET}"
+difft <(echo "$original_hashes") <(echo "$updated_hashes")
 
-# Update line numbers when the metadata section is on different lines
-metadata_begin_line=12
-metadata_end_line=16
-new_flake_file=$(mktemp)
-
-echo -e "\033[1;32m------ Updating flake.nix ------\033[0m"
-
-awk -v start=$metadata_begin_line -v end=$metadata_end_line -v newlines="$formatted_output" '
-NR == start {
-    print newlines
-    next
-}
-NR > start && NR <= end {
-    next
-}
-{
-    print
-}' flake.nix > "$new_flake_file" && mv "$new_flake_file" flake.nix
-
-echo -e "\033[1;32m------ The text above has been written to flake.nix ------\033[0m"
+# Asking for confirmation
+printf "Apply the changes above to flake.nix? [y/N]"
+read response
+if [[ "$response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+  # Update flake.nix
+  new_flake_file=$(mktemp)
+  awk -v start=$metadata_begin_line -v end=$metadata_end_line -v newlines="$updated_hashes" '
+  NR == start {
+      print newlines
+      next
+  }
+  NR > start && NR <= end {
+      next
+  }
+  {
+      print
+  }' flake.nix > "$new_flake_file" && mv "$new_flake_file" flake.nix
+  echo -e "${GREEN}Updated flake.nix${RESET}"
+else
+  echo -e "${YELLOW}Aborted without any change to flake.nix${RESET}"
+fi
